Move old tests to separate file
